% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statnet-package.R
\docType{package}
\name{statnet-package}
\alias{statnet-package}
\alias{statnet}
\title{Easily Install and Load the \code{statnet} Packages for Statistical Network
Analysis}
\description{
\code{statnet} is a collection of software packages for statistical network
analysis that are designed to work together, with a common data structure
and API, to provide seamless access to a broad range of network analytic and
graphical methodology.  This package is designed to make it easy to install
and load multiple \code{statnet} packages in a single step.
}
\details{
\code{statnet} software implements advances in network modeling based on
exponential-family random graph models (ERGM), as well as latent space
models and more traditional descriptive network methods. Together,
the set of packages provide a comprehensive framework for "tie-based" 
network analysis:  analyzing the distribution of ties in cross-sectional 
and dynamic networks.  There are tools for description, visualization
model estimation, model evaluation, and model-based network simulation.  
The statistical estimation and simulation functions
are based on a central Markov chain Monte Carlo (MCMC) algorithm that
has been optimized for speed and robustness.

The code is actively developed and maintained by the \code{statnet}
development team. New functionality is being added over time.

\code{statnet} packages are written in a combination of and \code{C} It is
can be used interactively from within the graphical user interface via a
command line, or in non-interactive (or ``batch'') mode to
allow longer or multiple tasks to be processed without user interaction.
The Statnet project uses an
open development process for the packages, hosted
on GitHub \url{https://github.com/statnet}, and contributions can be made
via pull requests.  Current versions of the packages are
available on the Comprehensive \R Archive Network (CRAN) at
\url{https://www.r-project.org/}.

Extensive workshop and training materials are also available online, 
please see the \code{statnet} project
website at \url{https://www.statnet.org/} for more information.

The full suite of packages has the following components (those automatically
downloaded with the \pkg{statnet} package are noted):

For data handling:

    (\code{\link[tergm]{tergm}}),
    (\code{\link[tergm]{simulate}}\code{},
    \code{\link[tergm]{mcmc.diagnostics}} and \code{\link[tergm]{gof}}).
    with two different types of discrete temporal network data:
}
\author{
Mark S. Handcock \email{handcock@stat.ucla.edu},\cr David R. Hunter
\email{dhunter@stat.psu.edu},\cr Carter T. Butts
\email{buttsc@uci.edu},\cr Steven M. Goodreau \email{goodreau@uw.edu},\cr
Pavel N. Krivitsky \email{pavel@uow.edu.au}, \cr Skye Bender-deMoll
\email{skyebend@skyeome.net}, \cr Samuel Jenness (for EpiModel)
\email{samuel.m.jenness@emory.edu}, and \cr Martina Morris
\email{morrism@uw.edu}

Maintainer: Martina Morris \email{morris@uw.edu}
}
