% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_statnet.R
\name{update_statnet}
\alias{update_statnet}
\title{Update the Component Packages of the Statnet Suite}
\usage{
update_statnet(..., ask = FALSE, checkBuilt = TRUE, addURLs = character())
}
\arguments{
\item{\dots}{Additional arguments to be passed to
\code{\link{update.packages}}.}

\item{ask, checkBuilt}{Arguments to \code{\link{update.packages}}
documentation. The defaults are different from those of that function.}

\item{addURLs}{Optional repository URLs in addition to CRAN, such as
\url{http://statnet.csde.washington.edu/preview}. Defaults to none.}
}
\value{
\code{update_statnet} returns NULL invisibly.
}
\description{
A wrapper around \code{\link{update.packages}} to update the component
packages of Statnet Suite to their latest versions.
}
\details{
Updates the list component packages of Statnet Suite, using
\code{\link{setRepositories}} and \code{\link{update.packages}}.

Since there are no good ways to update packages once they are loaded, this
function should be called immediately after restarting R.
}
\examples{

\dontrun{
# Update from CRAN
statnet::update_statnet()

# Update from statnet.org's preview release, taking packages from CRAN
# as needed
statnet::update_statnet(addURLs="http://statnet.csde.washington.edu/preview")
}

}
\seealso{
\code{\link{setRepositories}}, \code{\link{update.packages}},
\code{\link{install.packages}}
}
\keyword{utilities}
