\name{update_statnet}
\alias{update_statnet}
\title{Updates the Component Packages of statnet}
\usage{
update_statnet(\dots)
}
\arguments{
  \item{\dots}{
    Arguments passed to \code{\link{update.packages}}
  }
}
\description{
  
  This function has been deprecated, and may be removed in a future
  release. Currently, it exits with an error, advising the user to use
  \code{\link{update.packages}} instead. For installing public beta
  releases from the Statnet Project reposititories, see
  \url{http://www.statnet.org/preview}.
  
}
\value{
  \code{update_statnet} returns NULL invisibly.  
}
\seealso{
  \code{\link{update.packages}}, \code{\link{install.packages}}
}
\keyword{utilities}
